// poc/coo-web/src/api/generated/personalize/models/TimelineCursor.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { TimelineActivity } from './TimelineActivity';
import {
    TimelineActivityFromJSON,
    TimelineActivityFromJSONTyped,
    TimelineActivityToJSON,
    TimelineActivityToJSONTyped,
} from './TimelineActivity';

/**
 * 
 * @export
 * @interface TimelineCursor
 */
export interface TimelineCursor {
    /**
     * 
     * @type {Array<TimelineActivity>}
     * @memberof TimelineCursor
     */
    activities?: Array<TimelineActivity>;
    /**
     * 
     * @type {string}
     * @memberof TimelineCursor
     */
    nextCursor?: string;
    /**
     * 
     * @type {boolean}
     * @memberof TimelineCursor
     */
    hasMore?: boolean;
}

/**
 * Check if a given object implements the TimelineCursor interface.
 */
export function instanceOfTimelineCursor(value: object): value is TimelineCursor {
    return true;
}

export function TimelineCursorFromJSON(json: any): TimelineCursor {
    return TimelineCursorFromJSONTyped(json, false);
}

export function TimelineCursorFromJSONTyped(json: any, ignoreDiscriminator: boolean): TimelineCursor {
    if (json == null) {
        return json;
    }
    return {
        
        'activities': json['activities'] == null ? undefined : ((json['activities'] as Array<any>).map(TimelineActivityFromJSON)),
        'nextCursor': json['nextCursor'] == null ? undefined : json['nextCursor'],
        'hasMore': json['hasMore'] == null ? undefined : json['hasMore'],
    };
}

export function TimelineCursorToJSON(json: any): TimelineCursor {
    return TimelineCursorToJSONTyped(json, false);
}

export function TimelineCursorToJSONTyped(value?: TimelineCursor | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'activities': value['activities'] == null ? undefined : ((value['activities'] as Array<any>).map(TimelineActivityToJSON)),
        'nextCursor': value['nextCursor'],
        'hasMore': value['hasMore'],
    };
}