// poc/coo-web/src/api/generated/statistics/models/Statistics.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Statistics
 */
export interface Statistics {
    /**
     * 
     * @type {string}
     * @memberof Statistics
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof Statistics
     */
    subjectId?: string;
    /**
     * 
     * @type {string}
     * @memberof Statistics
     */
    subjectType?: StatisticsSubjectTypeEnum;
    /**
     * 
     * @type {number}
     * @memberof Statistics
     */
    repliesCount?: number;
    /**
     * 
     * @type {number}
     * @memberof Statistics
     */
    likesCount?: number;
    /**
     * 
     * @type {number}
     * @memberof Statistics
     */
    viewsCount?: number;
    /**
     * 
     * @type {Date}
     * @memberof Statistics
     */
    updatedAt?: Date;
}


/**
 * @export
 */
export const StatisticsSubjectTypeEnum = {
    Coo: 'COO',
    Reply: 'REPLY'
} as const;
export type StatisticsSubjectTypeEnum = typeof StatisticsSubjectTypeEnum[keyof typeof StatisticsSubjectTypeEnum];


/**
 * Check if a given object implements the Statistics interface.
 */
export function instanceOfStatistics(value: object): value is Statistics {
    return true;
}

export function StatisticsFromJSON(json: any): Statistics {
    return StatisticsFromJSONTyped(json, false);
}

export function StatisticsFromJSONTyped(json: any, ignoreDiscriminator: boolean): Statistics {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'subjectId': json['subjectId'] == null ? undefined : json['subjectId'],
        'subjectType': json['subjectType'] == null ? undefined : json['subjectType'],
        'repliesCount': json['repliesCount'] == null ? undefined : json['repliesCount'],
        'likesCount': json['likesCount'] == null ? undefined : json['likesCount'],
        'viewsCount': json['viewsCount'] == null ? undefined : json['viewsCount'],
        'updatedAt': json['updatedAt'] == null ? undefined : (new Date(json['updatedAt'])),
    };
}

export function StatisticsToJSON(json: any): Statistics {
    return StatisticsToJSONTyped(json, false);
}

export function StatisticsToJSONTyped(value?: Statistics | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'subjectId': value['subjectId'],
        'subjectType': value['subjectType'],
        'repliesCount': value['repliesCount'],
        'likesCount': value['likesCount'],
        'viewsCount': value['viewsCount'],
        'updatedAt': value['updatedAt'] == null ? undefined : ((value['updatedAt']).toISOString()),
    };
}